Used scrape to get a book list
I'm thinking that perhaps it's best to just use the database Alex found and see if it has summaries as well. Lining up the books might be challenging. There are a lot of duplicates.

On the bright side, scraping with scrapy is super easy once you figure out how to do it.